♻️ Rename the module as implementations for prompt provider.
